+++ docs/misc/changelog.rst
@@ -70,7 +70,7 @@ Others:
 - Removed ignores on Flake8 F401 (@ZikangXiong)
 - Upgraded GitHub CI/setup-python to v4 and checkout to v3
 - Set tensors construction directly on the device
-- Standardized the use of ``from gym import spaces``
+- Standardized the use of ``from gymnasium import spaces``
 
 Release 1.6.2 (2022-10-10)
 --------------------------